DOC: Specify py2k-compatible version in py3k statement
The `python>=3.6` requirement at the top of the file dates to early 2020, before v0.1 of MONETIO.  Before this, MONET and MONETIO were one package.  It looks like the last commit in MONET 2.1.5 merges in the work to split MONET and MONETIO, so someone wanting MONETIO functionality from MONET would need a MONET version prior to that.